requirements.txt
torch==2.4.1
torchvision==0.19.1+cu121
torchaudio==2.4.1
transformers==4.46.3
tokenizers==0.20.3
peft==0.13.2
datasets==3.1.0
evaluate==0.4.6
PyYAML==6.0.1
tqdm==4.67.1
numpy==1.21.5
pandas==1.2.4
scikit-learn==1.2.2
huggingface-hub==0.35.3
safetensors==0.5.3
accelerate==1.0.1
tensorboard==2.12.0
matplotlib==3.5.1
fsspec==2023.6.0
requests==2.32.4
